+++ datalad_container/tests/test_run.py
@@ -16,7 +16,8 @@ testimg_url = 'shub://datalad/datalad-container:testhelper'
 
 
 @with_tempfile
-def test_container_files(path):
+@with_tempfile
+def test_container_files(path, super_path):
     ds = Dataset(path).create()
     # plug in a proper singularity image
     ds.containers_add(
@@ -32,6 +33,7 @@ def test_container_files(path):
         name='mycontainer',
         updateurl=testimg_url)
     ok_clean_git(path)
+
     # now we can run stuff in the container
     # and because there is just one, we don't even have to name the container
     res = ds.containers_run(['dir'] if on_windows else ['ls'])
@@ -42,3 +44,41 @@ def test_container_files(path):
     # this command changed nothing
     assert_result_count(
         res, 1, action='add', status='notneeded', path=ds.path, type='dataset')
+
+    # same thing as we specify the container by its name:
+    res = ds.containers_run(['dir'] if on_windows else ['ls'],
+                            container_name='mycontainer')
+    # container becomes an 'input' for `run` -> get request, but "notneeded"
+    assert_result_count(
+        res, 1, action='get', status='notneeded',
+        path=op.join(ds.path, 'righthere'), type='file')
+    # this command changed nothing
+    assert_result_count(
+        res, 1, action='add', status='notneeded', path=ds.path, type='dataset')
+
+    # we can also specify the container by its path:
+    res = ds.containers_run(['dir'] if on_windows else ['ls'],
+                            container_name=op.join(ds.path, 'righthere'))
+    # container becomes an 'input' for `run` -> get request, but "notneeded"
+    assert_result_count(
+        res, 1, action='get', status='notneeded',
+        path=op.join(ds.path, 'righthere'), type='file')
+    # this command changed nothing
+    assert_result_count(
+        res, 1, action='add', status='notneeded', path=ds.path, type='dataset')
+
+    # Now, test the same thing, but with this dataset being a subdataset of
+    # another one:
+
+    super_ds = Dataset(super_path).create()
+    super_ds.install("sub", source=path)
+
+    res = super_ds.containers_run(['dir'] if on_windows else ['ls'])
+    # container becomes an 'input' for `run` -> get request (needed this time)
+    assert_result_count(
+        res, 1, action='get', status='ok',
+        path=op.join(super_ds.path, 'sub', 'righthere'), type='file')
+    # this command changed nothing
+    assert_result_count(
+        res, 1, action='add', status='notneeded', path=super_ds.path, type='dataset')
+